Implement caching for CSV maker to improve performance
Recently we saw it take ~10 minutes to generate the CSV, due to all of the
string searches for make/model in the HTML. This was unfortunate because
this step has to run on all pages every time there are new pages.

Now the CSV row for each file is cached in its own file. We only need to
do the slow string searches for new pages. (Or the cache can be cleared in
case the CSV format changes.)