Be smart about scraping
We introduce a concept of page sources (RSS URLs) -- each URL can be for
a different subdomain or include different search terms, etc.
Each source is scored based on how many new pages it is expected to
provide, and we always hit the highest-scoring source. Score also accounts
for data staleness.